requirements.txt
click==8.1.3
csscompressor==0.9.5
htmlmin==0.1.12
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
pathtools==0.1.2
ply==3.11
PyYAML==6.0
slimit==0.8.1
uuid==1.30
watchdog==2.1.9
webptools==0.0.9